requirements.txt
oyaml
GitPython
colorama
jinja2
passlib
requests
pytest
google-cloud-firestore